Implemented Services and Extensions
Implemented Services responsible for the API calls on the client side.
The Services is using a UnitOfWork-like Style with a generic Service class and unique subclasses attached to the parent class.

Also added an extension class library with an extension method for TEntity to manipulate the name of the type.