Added DTO's, SoldTicketModel, and fixed Repositories Missing Asyncs
Created a DTO folder in the shared project, and added some DTO's mirroring the Models and ViewModels.

Fixed Some of the missing Async Methods in the UnitOfWork.

Added a SoldTicketModel + ViewModel + DTO

Added a SoldTicketsControlller and made a draft of a HTTPGet method.

Removed Constructors from ViewModels that caused an error.

Added a constructor in the CountryViewModel for the SoldTicketController method.